% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{add_spline_threshold}
\alias{add_spline_threshold}
\title{Use spline lookup to restrict `ObservedExpectedGrid` to potential clusters}
\usage{
add_spline_threshold(oe_grid, spline_lookup = NULL)
}
\arguments{
\item{oe_grid}{An object of class `ObservedExpectedGrid` generated by
\code{generate_observed_expected()}}

\item{spline_lookup}{default NULL; either a spline lookup table, which is a
data frame that has at least two columns: including "observed" and
"spl_thresh", OR a string indicating to use one of the built in lookup
tables: i.e. one of \code{"001", "005", "01", "05"}. If NULL, the default
table will be 01 (i.e. \code{spline_01} dataset)}
}
\value{
an object of class `ClusterAlertTable` which is simply a data frame
  containing rows of the input `oe_grid`` that represent the reduced set of
  candidate alert clusters
}
\description{
Function takes a spline lookup table (or uses package default), and an object
of class `ObservedExpectedGrid` and identifies which rows in each potential
centroid have observed over expected values that exceed a threshold for that
observed value
}
\examples{
case_grid <- generate_case_grids(
  example_count_data, example_count_data[, max(date)]
)
nci <- gen_nearby_case_info(
  cg = case_grid,
  distance_matrix = county_distance_matrix("OH")[["distance_matrix"]],
  distance_limit = 25
)
obs_exp_grid <- generate_observed_expected(
  nearby_counts = nci,
  case_grid = case_grid
)
add_spline_threshold(oe_grid = obs_exp_grid)
add_spline_threshold(oe_grid = obs_exp_grid, spline_lookup = "01")
}
