% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{check_vars}
\alias{check_vars}
\title{check for variables in frame}
\usage{
check_vars(d, required)
}
\arguments{
\item{d}{input data frame to check for variables}

\item{required}{vector of column names that must be in `d`}
}
\description{
Function checks for variables in frame
}
