% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{compress_clusters}
\alias{compress_clusters}
\title{Compress a cluster_alert_table}
\usage{
compress_clusters(cluster_alert_table, distance_matrix)
}
\arguments{
\item{cluster_alert_table}{an object of class `ClusterAlertTable`}

\item{distance_matrix}{a distance matrix with named dimensions, where the
names are the candidate locations used to generate ClusterAlertTable}
}
\value{
a list including a data.frame of clusters and another frame of
  individual location counts
}
\description{
Function reduces an object of class ClusterAlertTable to the final set of
clusters and locations. The idea of this function is to retain only the most
significant, non-overlapping clusters from the cluster alert table. The
surrogate for significance is 'alertGap', or log(observed/expected) minus the
threshold that the spline assigns to the observed value`. The logic below
keeps two running tables, the table 'dt_keep' of clusters to be kept, in
descending order of 'alertGap', and 'dt_clust', the remaining rows of the
cluster alert table, which are reduced each time a cluster is accepted into
dt_keep Each row of the cluster alert table represents a candidate cluster,
with a column 'target', the cluster center, and a column 'location', the most
distant location from the center. Each time a cluster is accepted into
dt_keep, jClust is incremented, and remaining rows of 'dt_clust' are dropped
if either 'target' or 'location' is the center of the newly accepted cluster
in 'dt_keep'
}
