% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{County Location Dataset}
\format{
A data frame with 3,144 rows and 6 columns:
\describe{
  \item{state_name, state}{full and abbreviated names for states}
  \item{state_fips, fips}{state and county fips codes}
  \item{longitude, latitude}{numeric coordinates for fips}
}
}
\source{
`tigris` package
}
\usage{
counties
}
\description{
A data set that provides latitude and longitude for each
county in the United Sates
}
\keyword{datasets}
