% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{county_distance_matrix}
\alias{county_distance_matrix}
\title{Get distance matrix for counties within a state}
\usage{
county_distance_matrix(
  st,
  unit = c("miles", "kilometers", "meters"),
  source = c("tigris", "rnssp")
)
}
\arguments{
\item{st}{vector of 2 character strings denoting a state (or set of states),
or "US". If "US", then this is equivalent to calling
\code{us_distance_matrix()}.}

\item{unit}{string, one of "miles" (default), "kilometers", or "meters".
Indicating the desired unit for the distances}

\item{source}{string indicating either "tigris" (default) or "rnssp". Both
are built-in datasets (i.e. are part of this package). The default
("tigris") uses county names and locations as found in tigris 2024. The
"rnssp" option uses a package-stored version of the publicly available
shape file for counties from Rnssp package at
https://cdcgov.github.io/Rnssp/}
}
\value{
a named list of length two; first element (`loc_vec`) is a vector of
  locations and the second element (`distance_matrix`) is a square matrix
  containing the pairwise distance (in the given `unit`) between all
  locations.
}
\description{
Function returns a list of counties and a matrix with the distance between
those counties. leverages a built in dataset (`counties`).
}
\examples{
county_distance_matrix("MD", source = "tigris")
county_distance_matrix("WI", source = "rnssp", unit = "kilometers")
}
