% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{create_custom_dist_list}
\alias{create_custom_dist_list}
\title{Create a sparse distance list from custom location data}
\usage{
create_custom_dist_list(
  df,
  label_var,
  lat_var,
  long_var,
  threshold,
  unit = c("miles", "kilometers", "meters")
)
}
\arguments{
\item{df}{data.frame containing label and coordinate columns}

\item{label_var}{character scalar; column name used as location label
(must be unique and non-missing)}

\item{lat_var}{character scalar; latitude column name.}

\item{long_var}{character scalar; longitude column name.}

\item{threshold}{numeric scalar distance cutoff in units of \code{unit}}

\item{unit}{string, one of "miles" (default), "kilometers", or "meters"}
}
\value{
a named list, where each element, named by a target location, is a named
 vector of distances that are within `threshold` `units` of the target.
}
\description{
This function is a custom-data version of \code{create_dist_list()}.
It returns a list of named numeric vectors where each list element contains
only locations within \code{threshold} distance units of a target location.
}
\examples{
md <- tract_generator("MD")
dlist <- create_custom_dist_list(
  df = md,
  label_var = "geoid",
  lat_var = "latitude",
  long_var = "longitude",
  threshold = 15,
  unit = "miles"
)
}
