% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{.clean_up_alert_table}
\alias{.clean_up_alert_table}
\title{Clean and finalize cluster alert table:}
\usage{
.clean_up_alert_table(clt)
}
\arguments{
\item{clt}{cluster alert table holding the cluster specific information}
}
\value{
data.table of cleaned up column names
}
\description{
Function takes the raw cluster alert table, and cleans it up, namely
removing some columns that are not needed, renaming some columns, and
setting the final order of columns
}
\keyword{internal}
