% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{.distance_meters_from_coords}
\alias{.distance_meters_from_coords}
\title{Helper function gets the distance in meter between
pairs of coordinates. Note that \code{coords} must be a matrix
or frame, where the first col is longitude and the second
column is latitude}
\usage{
.distance_meters_from_coords(coords)
}
\arguments{
\item{coords}{Matrix-like object with longitude in column 1 and latitude in
column 2.}
}
\value{
Numeric square matrix of pairwise distances in meters.
}
\description{
Helper function gets the distance in meter between
pairs of coordinates. Note that \code{coords} must be a matrix
or frame, where the first col is longitude and the second
column is latitude
}
\keyword{internal}
