% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{.identify_states_from_counties}
\alias{.identify_states_from_counties}
\title{Identify states from county location-date-count frame}
\usage{
.identify_states_from_counties(df, location = "location")
}
\description{
Identify states from county location-date-count frame
}
\keyword{internal}
