% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{.identify_states_from_zips}
\alias{.identify_states_from_zips}
\title{Identify states from zipcode location-date-count frame}
\usage{
.identify_states_from_zips(df, location = "location")
}
\description{
Identify states from zipcode location-date-count frame
}
\keyword{internal}
