% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{.numeric_location_coords}
\alias{.numeric_location_coords}
\title{Helper function reduces a data frame to only those rows where latitude
and longitude are not missing}
\usage{
.numeric_location_coords(locs)
}
\arguments{
\item{locs}{A \code{data.table} with \code{latitude} and \code{longitude}
columns.}
}
\value{
A \code{data.table} filtered to non-missing latitude rows with
  numeric latitude/longitude columns.
}
\description{
Helper function reduces a data frame to only those rows where latitude
and longitude are not missing
}
\keyword{internal}
