% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{.resolve_coord_var_names}
\alias{.resolve_coord_var_names}
\title{Helper function resolves coordinate variable names}
\usage{
.resolve_coord_var_names(lat_var = NULL, long_var = NULL)
}
\arguments{
\item{lat_var}{Character scalar latitude column name or \code{NULL} to use
\code{"latitude"}.}

\item{long_var}{Character scalar longitude column name or \code{NULL} to use
\code{"longitude"}.}
}
\value{
A named list with elements \code{lat_var} and \code{long_var}.
}
\description{
Helper function resolves coordinate variable names
}
\keyword{internal}
