% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{.sparse_dist_list_from_locs}
\alias{.sparse_dist_list_from_locs}
\title{This is a helper function to create a named list of all the locations in
\code{locs} within \code{threshold_meters} of each loc in \code{locs}.}
\usage{
.sparse_dist_list_from_locs(locs, threshold_meters, meters_per_unit)
}
\arguments{
\item{locs}{A \code{data.table} with columns \code{location},
\code{latitude}, and \code{longitude}.}

\item{threshold_meters}{Numeric scalar distance threshold in meters.}

\item{meters_per_unit}{Numeric scalar conversion factor from output unit to
meters.}
}
\value{
Named list of numeric vectors of neighbor distances, keyed by
  location.
}
\description{
This is a helper function to create a named list of all the locations in
\code{locs} within \code{threshold_meters} of each loc in \code{locs}.
}
\keyword{internal}
