% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_count_data}
\alias{example_count_data}
\title{Example Count Dataset}
\format{
A data frame with 11,264 rows and 4 columns:
\describe{
  \item{location}{county FIPS code as character}
  \item{date}{date of observation}
  \item{count}{non-negative integer daily count}
}
}
\source{
Generated from a local development input using
`data-raw/build_example_count_data.R`
}
\usage{
example_count_data
}
\description{
Synthetic county-level example count data for package examples and tests. 
Generationg includes a in cases near chronological end to ensure there is
non-trivial cluster detection behavior.
}
\keyword{datasets}
