% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_ggplot_time_series}
\alias{generate_ggplot_time_series}
\title{Generate ggplot of timeseries}
\usage{
generate_ggplot_time_series(time_series_data, locations = "All Locations")
}
\arguments{
\item{time_series_data}{data frame generated by `generate_time_series_data`}

\item{locations}{vector of locations to limit to; default is "All Locations"}
}
\description{
Internal function to generate ggplot based timeseries
}
\keyword{internal}
