% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_heatmap}
\alias{generate_heatmap}
\title{Generate heatmap of data}
\usage{
generate_heatmap(heatmap_data, plot_type = c("ggplot", "plotly"), ...)
}
\arguments{
\item{heatmap_data}{data frame generated by `generate_heatmap_data`}

\item{plot_type}{string indicating either a "ggplot" or "plotly" result}

\item{...}{passed onto plotly}
}
\value{
a ggplot or plotly object
}
\description{
Generate a ggplot heatmap of count information by date and location given a
frame of count-by-location-and-date data.
}
