% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_heatmap_data}
\alias{generate_heatmap_data}
\title{Get heat map data from a set of location, date, count data}
\usage{
generate_heatmap_data(
  data,
  end_date = NULL,
  locations = NULL,
  baseline_length = 90,
  test_length = 7,
  guard = 0,
  break_points = c(-1, 2, 4, 9, 19, Inf),
  break_labels = c("0-1", "2-4", "5-9", "10-19", "20+")
)
}
\arguments{
\item{data}{data frame with (at least) three columns: location, date, count}

\item{end_date}{date indicating end of test interval; if not provided the
last date in `dt` will be used}

\item{locations}{a vector of locations to subset the table; if none provided
then all locations will be used}

\item{baseline_length}{numeric (default=90) number of days in baseline
interval}

\item{test_length}{numeric (default=7) number of days in test interval}

\item{guard}{numeric (default=0) number of days between baseline and test
interval}

\item{break_points}{break points for the discrete groups (default =
\code{c(-1,2,4,9,19,Inf)})}

\item{break_labels}{string vector of labels for the groups (default =
\code{c("0-1", "2-4", "5-9", "10-19", "20+")})}
}
\value{
a data frame of heat map data
}
\description{
Generate heat map data frame count information by date and location given an
input frame of count-by-location-and-date data.
}
