% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{generate_observed_expected}
\alias{generate_observed_expected}
\title{Generate the observed and expected information}
\usage{
generate_observed_expected(
  nearby_counts,
  case_grid,
  adjust = FALSE,
  adj_constant = 1
)
}
\arguments{
\item{nearby_counts}{an object of class `NearbyClusterGrids`}

\item{case_grid}{an object of class `CaseGrids`}

\item{adjust}{boolean default TRUE, set to \code{FALSE} to avoid adding one
to the expected when it is zero. Could result in errors.}

\item{adj_constant}{numeric (default=1.0); this is the constant to be added
if \code{baseline_adjustment == 'add_one'} or \code{baseline_adjustment ==
'add_one'}}
}
\value{
a dataframe of class `ObservedExpectedGrid`, which is simply a data
  frame with the observed and expected calculation incorporated for all
  target locations
}
\description{
Function takes an object of class `NearbyClusterGrids`, as returned from
\code{gen_nearby_case_info()}, and an object of class `CaseGrids`, as
returned from \code{generate_case_grids()}, and adds observed and expected
information.
}
\examples{
case_grid <- generate_case_grids(
  example_count_data,
  example_count_data[, max(date)]
)
nci <- gen_nearby_case_info(
  cg = case_grid,
  distance_matrix = county_distance_matrix("OH")[["distance_matrix"]],
  distance_limit = 25
)
generate_observed_expected(
  nearby_counts = nci,
  case_grid = case_grid
)
}
