% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_summary_table}
\alias{generate_summary_table}
\title{Summary count-by-location-and-date data, given baseline and test interval
lengths, and an end-date for the test interval}
\usage{
generate_summary_table(
  data,
  end_date = NULL,
  locations = NULL,
  baseline_length = 90,
  test_length = 7,
  guard = 0,
  cut_vec = c(0, 1.5, 2.5, 5.5, 10.5, Inf),
  cut_labels = c("Nr. Locs, daily mean 1 or less", "Nr. Locs, daily mean 2",
    "Nr. Locs, daily mean 3-5", "Nr. Locs, daily mean 6-10", "Nr. Locs, daily mean >10")
)
}
\arguments{
\item{data}{data frame with (at least) three columns: location, date, count}

\item{end_date}{date indicating end of test interval; if not provided the
last date in `dt` will be used}

\item{locations}{a vector of locations to subset the table; if none provided
then all locations will be used}

\item{baseline_length}{numeric (default=90) number of days in baseline
interval}

\item{test_length}{numeric (default=7) number of days in test interval}

\item{guard}{numeric (default=0) number of days between baseline and test
interval}

\item{cut_vec}{numeric vector of n cut points to examine categories of daily
mean counts}

\item{cut_labels}{character vector of labels for the n-1 categories created
by `cut_vec`}
}
\value{
data frame of summary statistics
}
\description{
Function will return a summary data frame of information related to a given
count-by-location-and-date dataset, provided the user gives the count data, a
set of locations, and the length of the baseline and test intervals, and and
end date for the test interval. Note that a guard, a buffer between the end
of the baseline interval and the test interval can be provided.
}
