% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_time_series_data}
\alias{generate_time_series_data}
\title{Generate time series data}
\usage{
generate_time_series_data(
  data,
  end_date = NULL,
  locations = NULL,
  baseline_length = 90,
  test_length = 7,
  guard = 0
)
}
\arguments{
\item{data}{data frame with (at least) three columns: location, date, count}

\item{end_date}{date indicating end of test interval; if not provided the
last date in `dt` will be used}

\item{locations}{a vector of locations to subset the table; if none provided
then all locations will be used}

\item{baseline_length}{numeric (default=90) number of days in baseline
interval}

\item{test_length}{numeric (default=7) number of days in test interval}

\item{guard}{numeric (default=0) number of days between baseline and test
interval}
}
\value{
a ggplot object
}
\description{
Function returns a time series of counts-by-location-and-date data,
given length of baseline and test intervals, and an end date for the
test-interval
}
