% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_time_series_plot}
\alias{generate_time_series_plot}
\title{Generate timeseries plot data}
\usage{
generate_time_series_plot(
  time_series_data,
  end_date = NULL,
  plot_type = c("ggplot", "plotly"),
  locations = "All Locations",
  ...
)
}
\arguments{
\item{time_series_data}{data frame generated by `generate_time_series_data`}

\item{end_date}{optional end date to truncate date}

\item{plot_type}{string indicating either a "ggplot" or "plotly" result}

\item{locations}{string indicating location name (defaults to "All Locations")}

\item{...}{passed onto plotly}
}
\value{
a ggplot or plotly object
}
\description{
Generate a timeseries plot of count information by date and location given a
frame of count-by-location-and-date data and an optional end_date
}
