% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{get_test_dates}
\alias{get_test_dates}
\title{Generate test dates vector}
\usage{
get_test_dates(end_date, test_length)
}
\arguments{
\item{end_date}{End date of the test interval}

\item{test_length}{(integer) length of the test interval in days}
}
\value{
vector of dates
}
\description{
Function to generate test dates given an end date and test length
}
