% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{ggplot_heatmap}
\alias{ggplot_heatmap}
\title{Generate ggplot heatmap}
\usage{
ggplot_heatmap(heatmap_data)
}
\arguments{
\item{heatmap_data}{data frame generated by `generate_heatmap_data`}
}
\description{
Internal function to generate ggplot based heatmap
}
\keyword{internal}
