% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsClusterDetect-package.R
\docType{package}
\name{gsClusterDetect-package}
\alias{gsClusterDetect}
\alias{gsClusterDetect-package}
\title{gsClusterDetect: Utilities for Geo-Spatial Cluster Detection and Significance Classification}
\description{
Provides utilities for manipulating time series of location-based counts of events to detect geo-spatial clusters. Significance of these clusters is determined using a set of models that classify based on a learned relationship between observed and the log(observed/expected) ratio of counts. The approach implemented here is similar to prospective space-time estimation of clusters using the scan statistic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lmullany/gsClusterDetect}
  \item \url{https://lmullany.github.io/gsClusterDetect/}
  \item Report bugs at \url{https://github.com/lmullany/gsClusterDetect/issues}
}

}
\author{
\strong{Maintainer}: Luke C. Mullany \email{luke.mullany@jhuapl.edu}

Authors:
\itemize{
  \item Luke C. Mullany \email{luke.mullany@jhuapl.edu}
  \item Howard S. Burkom
}

}
\keyword{internal}
