% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{identify_states}
\alias{identify_states}
\title{Identify states from location frame}
\usage{
identify_states(df, level = c("zip", "county"), location = "location")
}
\arguments{
\item{df}{frame/data.table that contains a column of locations}

\item{level}{string either "zip" (default) or "county"}

\item{location}{string column identifying the location (default is
"location")}
}
\value{
vector of state abbreviations
}
\description{
Helper function to identify states represented in a location frame; this is
helpful because the location functions are faster and more efficient when
states are specified.
}
\examples{
identify_states(example_count_data, level = "zip")
identify_states(example_count_data, level = "county")
}
