% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{plotly_heatmap}
\alias{plotly_heatmap}
\title{Generate plotly heatmap}
\usage{
plotly_heatmap(
  heatmap_data,
  x = "date",
  y = "location",
  z = "count",
  logscale = FALSE
)
}
\arguments{
\item{heatmap_data}{data frame generated by `generate_heatmap_data`}

\item{x}{name of date column, default is "date"}

\item{y}{name of location column, default is "y",}

\item{logscale}{boolean (default FALSE); set to TRUE to view on log scale}
}
\value{
a plotly object
}
\description{
Internal function to generate plotly based heatmap
}
\keyword{internal}
