% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{reduce_clusters_to_min}
\alias{reduce_clusters_to_min}
\title{Filter clusters on minimum overall count}
\usage{
reduce_clusters_to_min(cl, minimum = 0)
}
\arguments{
\item{cl}{a object of class \code{clusters}, as returned from
\code{compress_clusters}}

\item{minimum}{numeric (default = 0); minimum number across all locations in
a cluster in order to retain}
}
\value{
an object of class \code{clusters}
}
\description{
Function takes a set of clusters identified via \code{compress_clusters()}
and a minimum threshold for counts, and reduces the identified clusters to
only those clusters where the total number of observed across the cluster
meets that minimum threshold.
}
\examples{
cl <- find_clusters(
  cases = example_count_data,
  distance_matrix = county_distance_matrix("OH")[["distance_matrix"]],
  detect_date = example_count_data[, max(date)],
  distance_limit = 50
)
reduce_clusters_to_min(cl, 50)
}
