% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{sf_to_plotly_polygons}
\alias{sf_to_plotly_polygons}
\title{Convert sf to plotly polygons}
\usage{
sf_to_plotly_polygons(md)
}
\arguments{
\item{md}{list as returned by \code{prepare_map_data()}}
}
\description{
Helper function that takes the shape file merged with cluster information
as returned by the "map_data" element of \code{prepare_map_data()} and
converts each location to long polygon coordinate rows suitable for plotly.
}
\details{
This function only prepares drawing geometry. Hover information is prepared
separately in \code{prepare_map_data()} as the "hover_data" element.
}
\keyword{internal}
