% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spline_005}
\alias{spline_005}
\title{Spline Lookup Table - 0.005}
\format{
A data frame with 399 rows and 2 columns:
\describe{
  \item{observed}{number of observed in cluster}
  \item{spl_thresh}{log observed-over-expected above which
   cluster is significant at the 0.005 level}
}
}
\source{
package authors
}
\usage{
spline_005
}
\description{
Spline threshold lookup table, p-value = 0.005
}
\keyword{datasets}
