% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spline_05}
\alias{spline_05}
\title{Spline Lookup Table - 0.05}
\format{
A data frame with 399 rows and 2 columns:
\describe{
  \item{observed}{number of observed in cluster}
  \item{spl_thresh}{log observed-over-expected above which cluster is significant at the 0.05 level}
}
}
\source{
package authors
}
\usage{
spline_05
}
\description{
Spline threshold lookup table, p-value = 0.05
}
\keyword{datasets}
