% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{state_distance_matrix}
\alias{state_distance_matrix}
\title{Get distance matrix for states in the US}
\usage{
state_distance_matrix(unit = c("miles", "kilometers", "meters"))
}
\arguments{
\item{unit}{string, one of "miles" (default), "kilometers", or "meters".
Indicating the desired unit for the distances}
}
\value{
a named list of length two; first element (`loc_vec`) is a vector of
locations and the second element (`distance_matrix`) is a square matrix
containing the pairwise distance (in the given `unit`) between all locations.
}
\description{
Function returns a list of states and a matrix with the distance between
those states. leverages a built in dataset (`states`)
}
\examples{
state_distance_matrix()
state_distance_matrix(unit="kilometers")
}
