% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{State Location Dataset}
\format{
A data frame with 56 rows and 5 columns:
\describe{
  \item{state_name, state}{full and abbreviated names for states}
  \item{state_fips}{state fips codes}
  \item{longitude, latitude}{numeric coordinates for state fips}
}
}
\source{
`tigris` package
}
\usage{
states
}
\description{
A data set that provides latitude and longitude for each
state in the United Sates
}
\keyword{datasets}
