% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{validate_clusters}
\alias{validate_clusters}
\title{Validates cluster object}
\usage{
validate_clusters(cl)
}
\arguments{
\item{cl}{object to validate}
}
\description{
Function ingests an object and checks its validity as an object returned
from \code{find_clusters()}
}
\keyword{internal}
