% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{zip_distance_matrix}
\alias{zip_distance_matrix}
\title{Get distance matrix for zip codes within a state}
\usage{
zip_distance_matrix(st, unit = c("miles", "kilometers", "meters"))
}
\arguments{
\item{st}{two-character string denoting a state}

\item{unit}{string, one of "miles" (default), "kilometers", or "meters".
Indicating the desired unit for the distances}
}
\value{
a named list of length two; first element (`loc_vec`) is a vector of
locations and the second element (`distance_matrix`) is a square matrix
containing the pairwise distance (in the given `unit`) between all locations.
}
\description{
Function returns a list of zipcodes and a matrix with the distance between
those zip codes. leverages a built in dataset (`zipcodes`) that maps
zipcodes to counties.
}
\examples{
zip_distance_matrix("MD")
}
