% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{add_location_counts}
\alias{add_location_counts}
\title{Add location counts to cluster location list}
\usage{
add_location_counts(cluster_list, cases)
}
\arguments{
\item{cluster_list}{output list from 'compress_clusters', cluster summary
rows and locations in each cluster}

\item{cases}{original data in 3-column format of location, count, date}
}
\value{
the cluster list from compress_clusters with individual location
  counts appended
}
\description{
Add counts of individual cluster locations Operates on the output list of the
'compress_clusters()' component Calculates indiv. location counts for each
cluster, appends to the cluster location list
}
