% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{add_spline_threshold}
\alias{add_spline_threshold}
\title{Use spline lookup to restrict `ObserveExpectedGrid` to potential clusters}
\usage{
add_spline_threshold(oe_grid, spline_lookup)
}
\arguments{
\item{oe_grid}{An object of class `ObservedExpectedGrid` generated by
`generate_observed_expected()`}

\item{spline_lookup}{a spline lookup table; a data frame that has at least
two columns: including "observed" and "spl_thresh"}
}
\value{
a frame containing rows of oe_grid that are candidate alert clusters
}
\description{
Function takes a spline lookup table, and an object of class
ObservedExpectedGrid and identfies which rows in each potential centroid have
observed over expected values that exceed a threshold for that observed value
TODO: add a default system data object that contains a default spline lookup?
}
