% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{compress_clusters_fast}
\alias{compress_clusters_fast}
\title{Fast version of compress clusters}
\usage{
compress_clusters_fast(cluster_alert_table, distance_matrix)
}
\arguments{
\item{cluster_alert_table}{an object of class `ClusterAlertTable`}

\item{distance_matrix}{a distance matrix with named dimensions, where the
names are the candidate locations used to generate ClusterAlertTable}
}
\value{
a list including a data.frame of clusters and another frame of
  individual location counts
}
\description{
Function reduces an object of class ClusterAlertTable to the final set of
clusters and locations. The idea of this function is to retain only the most
significant, non-overlapping clusters from the cluster alert table. The
surrogate for significance is 'alertGap', or log(observed/expected) minus the
threshold that the spline assigns to the observed value`.
}
