% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{county_distance_matrix}
\alias{county_distance_matrix}
\title{Get distance matrix for counties within a state}
\usage{
county_distance_matrix(
  st,
  unit = c("miles", "kilometers", "meters"),
  source = c("built_in", "rnssp")
)
}
\arguments{
\item{st}{two-character string denoting a state, or "US". If "US", then this
is equivalent to calling \code{us_distance_matrix()}.}

\item{unit}{string, one of "miles" (default), "kilometers", or "meters".
Indicating the desired unit for the distances}

\item{source}{string indicating either "built_in" (default) or "rnssp".
The latter simply uses a package-stored version of the publicly available
shape file for counties from Rnssp package at https://cdcgov.github.io/Rnssp/}
}
\description{
Function returns a list of counties and a matrix with the distance between
those counties. leverages a built in dataset (`counties`).
}
\examples{
# example code
county_distance_matrix("MD", source = "built_in")
}
