% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{county_distance_matrix}
\alias{county_distance_matrix}
\title{Get distance matrix for counties within a state}
\usage{
county_distance_matrix(
  st,
  unit = c("miles", "kilometers", "meters"),
  source = c("rnssp", "tigris")
)
}
\arguments{
\item{st}{two-character string denoting a state, or "US". If "US", then this
is equivalent to calling the us_distance_matrix() function}

\item{unit}{string, one of "miles" (default), "kilometers", or "meters".
Indicating the desired unit for the distances}

\item{source}{string indicating either "rnssp" (default) or "tigris". The
former will use `Rnssp::county_sf` file and leverage sf_centroid() and
sf_distance() functions to produce the distance matrix. The latter will use
this package's default `counties` dataset and `geosphere` package to get
distance. The results are slightly different.}
}
\description{
Function returns a list of counties and a matrix with the distance between
those counties. leverages a built in dataset (`counties`).
}
\examples{
# example code
county_distance_matrix("MD", source = "tigris")
county_distance_matrix("CT", "kilometers", source = "tigris")
}
