% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{create_dist_list}
\alias{create_dist_list}
\title{Generalized distance list as sparse list}
\usage{
create_dist_list(
  level = c("county", "zip"),
  threshold = {
     if (level == "county") 
         50
     else 15
 },
  st = NULL,
  unit = c("miles", "kilometers", "meters")
)
}
\arguments{
\item{level}{string either "county" (default) or "zip"}

\item{threshold}{numeric value; include in each location-specific named
vector only those locations that a within `threshold` distance units of the
target. The defaults is 50 (miles) and 15 (miles) for county and zip, respectively, but
these can be adjusted. Note if a different unit other than miles is used, then the
user should also adjust this parameter appropriately}

\item{st}{string; optional to specify a state; if NULL distances are returned
for all zip codes or counties in the US}

\item{unit}{string one of miles (default), kilometers, or meters; this is the
unit relevant to the threshold}
}
\description{
This function is an alternative to the package functions that create a square
distance matrix of dimension N, with all pairwise distances. In this approach
a list of named vectors is returned, where there is one element in the list
for each location, and each named vector holds the distance within
`threshold` of the location.
}
