% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{generate_observed_expected}
\alias{generate_observed_expected}
\title{Generate the observed and expected information}
\usage{
generate_observed_expected(
  nearby_counts,
  cases_grids,
  adjust = FALSE,
  adj_constant = 1
)
}
\arguments{
\item{nearby_counts}{an object of class `NearbyClusterGrids`}

\item{cases_grids}{an object of class `CaseGrids`}

\item{adjust}{boolean default TRUE, set to \code{FALSE} to avoid adding
one to the expected when it is zero. Could result in errors.}

\item{adj_contant}{numeric (default=1.0); this is the constant to be added if
\code{baseline_adjustment == 'add_one_global'} or \code{baseline_adjustment
== 'add_one'}}
}
\value{
a dataframe of class `ObservedExpectedGrid`
}
\description{
Function takes an object of class `NearbyClusterGrids`, as returned from the
`generate_nearby_case_information()`, and adds the observed and expected
information
}
