% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_functions.R
\name{get_nearby_locations}
\alias{get_nearby_locations}
\title{Add data counts for injected clusters}
\usage{
get_nearby_locations(center_location, distance_matrix, radius_miles)
}
\arguments{
\item{center_location}{location}

\item{distance_matrix}{a distance matrix}

\item{radius_miles}{a numeric value >0}
}
\value{
a data.table
}
\description{
Returns a 2-column table of locations near center location and associated
case counts
}
\examples{
dm <- zip_distance_matrix("MD")$distance_matrix
nearby_locations <- get_nearby_locations("21228", dm, 10)
}
