% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineClusterDetector-package.R
\docType{package}
\name{splineClusterDetector-package}
\alias{splineClusterDetector}
\alias{splineClusterDetector-package}
\title{splineClusterDetector: Utilities for cluster detection and significance classification using splines}
\description{
This package provides utilities for manipulating time series of location-based counts of events to detect clusters, and to determine the significance of these clusters via spline approach.
}
\keyword{internal}
