% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_functions.R
\name{st_injects}
\alias{st_injects}
\title{Add data counts for parameterized injected clusters}
\usage{
st_injects(
  cases,
  distance_matrix,
  target_loc,
  center_decile,
  radius_miles,
  nr_cases,
  nr_days,
  end_date
)
}
\arguments{
\item{cases}{data frame of cases}

\item{distance_matrix}{a distance matrix}

\item{target_loc}{a location into which the injection should occur}

\item{center_decile}{an integer value between 1 and 10, inclusive}

\item{radius_miles}{a numeric value >0}

\item{nr_cases}{number of cases to inject}

\item{nr_days}{number of days over which we want to inject cases}

\item{end_date}{last date of injection}
}
\value{
a two-element list; each element is a dataframe. The first is the
  full dataset with injected cases and the second is the injected cases only
}
\description{
Function st_injects returns a list of two objects 1. a full dataset as a
data.table with inject counts added according to design parameters. 2. a
table of only the inject counts, locations, and dates.
}
\examples{
\dontrun{
scen1 <- st_injects(
  cases, distance_matrix, target_loc, center_decile,
  radius_miles, nr_cases, nr_days, end_date
)
}
}
