% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{us_distance_matrix}
\alias{us_distance_matrix}
\title{Get distance matrix for all counties in the US}
\usage{
us_distance_matrix(unit = c("miles", "kilometers", "meters"))
}
\arguments{
\item{unit}{string, one of "miles" (default), "kilometers", or "meters".
Indicating the desired unit for the distances}
}
\description{
Function returns a list of counties and a matrix with the distance between
those counties. leverages a built in dataset (`counties`). Note that the
generation of this matrix can take a few seconds
}
\examples{
# example code
us_distance_matrix()
}
