% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{zip_distance_matrix}
\alias{zip_distance_matrix}
\title{Get distance matrix for zip codes within a state}
\usage{
zip_distance_matrix(st, unit = c("miles", "kilometers", "meters"))
}
\arguments{
\item{st}{two-character string denoting a state}

\item{unit}{string, one of "miles" (default), "kilometers", or "meters".
Indicating the desired unit for the distances}
}
\description{
Function returns a list of zipcodes and a matrix with the distance between
those zip codes. leverages a built in dataset (`zipcodes`) that maps
zipcodes to counties.
}
\examples{
# example code
zip_distance_matrix("MD")
zip_distance_matrix("TX", "kilometers")
}
