% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{Zipcode Location Dataset}
\format{
A data frame with 42,482 rows and 11 columns:
\describe{
  \item{id}{serial integer id (1, 2, 3, .. etc)}
  \item{zip_code}{5 digit string for zipcode}
  \item{state}{state abbreviation}
  \item{county}{county name}
  \item{region}{region name}
  \item{region_id}{id for region}
  \item{region_name}{region name}
  \item{pop, modified}{undocumented}
  \item{latitude, longitude}{numeric coordinates for zipcode}
}
}
\source{
unknown
}
\usage{
zipcodes
}
\description{
A data set that provides latitude and longitude for each
zipcode in the United Sates
}
\keyword{datasets}
